requirements.txt
alembic==0.8.8
bcrypt==3.1.1
boto==2.42.0
cffi==1.8.3
click==6.6
decorator==4.0.10
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Login==0.3.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
pbr==1.10.0
pycparser==2.14
python-editor==1.0.1
six==1.10.0
SQLAlchemy==1.0.15
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
Werkzeug==0.11.11
WTForms==2.1
